refactor: rename support cypress types
