+++ tests/component/cypress/support/index.ts
@@ -2,7 +2,7 @@
 // https://on.cypress.io/configuration
 
 import './commands'
-import './types'
+import './cy-components'
 
 Cypress.on('window:before:load', (win: any) => {
   win.handleFromCypress = function (request: {
